+++ spec/lib/karafka/cli/console_spec.rb
@@ -8,23 +8,43 @@ RSpec.describe_current do
   specify { expect(described_class).to be < Karafka::Cli::Base }
 
   describe '#call' do
-    let(:cmd) do
-      envs = [
-        "IRBRC='#{Karafka.gem_root}/.console_irbrc'",
-        'KARAFKA_CONSOLE=true'
-      ]
-      "#{envs.join(' ')} bundle exec irb -r #{Karafka.boot_file}"
+    context 'when running without rails' do
+      let(:cmd) do
+        envs = [
+          'KARAFKA_CONSOLE=true',
+          "IRBRC='#{Karafka.gem_root}/.console_irbrc'"
+        ]
+        "#{envs.join(' ')} bundle exec irb -r #{Karafka.boot_file}"
+      end
+
+      before do
+        allow(cli).to receive(:info)
+        allow(console_cli).to receive(:exec)
+      end
+
+      it 'expect to execute irb with boot file required' do
+        console_cli.call
+        expect(cli).to have_received(:info)
+        expect(console_cli).to have_received(:exec).with(cmd)
+      end
     end
 
-    before do
-      allow(cli).to receive(:info)
-      allow(console_cli).to receive(:exec)
-    end
+    context 'when running with rails' do
+      let(:cmd) do
+        'KARAFKA_CONSOLE=true bundle exec rails console'
+      end
 
-    it 'expect to execute irb with boot file required' do
-      console_cli.call
-      expect(cli).to have_received(:info)
-      expect(console_cli).to have_received(:exec).with(cmd)
+      before do
+        allow(::Karafka).to receive(:rails?).and_return(true)
+        allow(cli).to receive(:info)
+        allow(console_cli).to receive(:exec)
+      end
+
+      it 'expect to execute rails console' do
+        console_cli.call
+        expect(cli).to have_received(:info)
+        expect(console_cli).to have_received(:exec).with(cmd)
+      end
     end
   end
 end